requirements.txt
Flask==2.2.2
matplotlib==3.4.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
